missile/rocket behavior tweaks
changelog:
rocket explosion radius now scales off of projectile scale
added velocity modifier to rockets and the ability to set it through the bp_weapon_component
added homing behavior to missiles
current missile behavior: homes into closest target to the projectile

intended behavior: homes into a target highlighted near the player targeting reticle?
would have to develop additional targeting system, projects a cone in front of the player at all times and draws distance to a ray projected from the player (targeting reticle) and highlights closest target to the reticle, and directs that as target
(this is likely outside of the scope of this project, as such missiles will be cut for now. emphasis of the project will focus on rockets and guns for now and we'll revisit homing missiles later if there's time)